feature: Add prefix iterator.
